refactor legacy config stuff
