+++ apps/lsu/src/main/java/dev/candycup/lifestealutils/features/ah/AhSearchAutomation.java
@@ -1,6 +1,17 @@
 package dev.candycup.lifestealutils.features.ah;
 
+import dev.candycup.configura.serial.SerialEntry;
+import dev.candycup.lifestealutils.config.configurables.ConfigurableBoolean;
+import dev.candycup.lifestealutils.config.configurables.RequiresGaia;
+import lombok.Getter;
+
 public final class AhSearchAutomation {
+   @Getter
+   @SerialEntry(comment = "Enable the custom auction house interface overlay GUI")
+   @RequiresGaia(forceStateWhenDenied = "false")
+   @ConfigurableBoolean(location = "experiments.auctionui.experimentalauctionui", icon = "gold_ingot")
+   private static boolean customAhInterfaceEnabled = false;
+
    private static String pendingQuery;
    private static long pendingQueryExpiresAtMs;
    private static String activeQuery;
