+++ apps/lsu/src/main/java/dev/candycup/lifestealutils/features/alliances/AllianceNameDecorator.java
@@ -1,9 +1,13 @@
 package dev.candycup.lifestealutils.features.alliances;
 
-import dev.candycup.lifestealutils.Config;
+import dev.candycup.configura.serial.SerialEntry;
+import dev.candycup.lifestealutils.config.configurables.ConfigurableBoolean;
+import dev.candycup.lifestealutils.config.configurables.IncludeInAccordion;
 import dev.candycup.lifestealutils.event.LifestealUtilsEvents;
 import dev.candycup.lifestealutils.event.LifestealUtilsEvents.PlayerNameRenderEvent;
 import dev.candycup.lifestealutils.interapi.MessagingUtils;
+import lombok.Getter;
+import lombok.Setter;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
@@ -12,12 +16,43 @@ import net.minecraft.network.chat.contents.PlainTextContents;
 import java.util.List;
 
 public final class AllianceNameDecorator {
+    @Getter
+    @Setter
+    @SerialEntry(comment = "Whether to enable alliance features such as colored name tags.")
+    @ConfigurableBoolean(location = "alliances.general.enablealliances")
+    private static boolean enableAlliances = true;
+
+    @Getter
+    @Setter
+    @SerialEntry(comment = "Whether to render alliance prefixes in nametags")
+    @ConfigurableBoolean(location = "alliances.general.allianceprefixenabled")
+    private static boolean allianceNamePrefixEnabled = true;
+
+    @Getter
+    @Setter
+    @SerialEntry(comment = "Whether to apply list-specific alliance name colors to player names")
+    @ConfigurableBoolean(location = "alliances.general.alliancenamecolorenabled")
+    private static boolean allianceNameColorEnabled = true;
+
+    @Getter
+    @Setter
+    @SerialEntry(comment = "Allow alliance members to bypass ghosted chat filtering")
+    @ConfigurableBoolean(location = "alliances.general.allowalliancebypassghostedchat")
+    private static boolean allowAllianceBypassGhostedChat = true;
+
+    @Getter
+    @Setter
+    @SerialEntry(comment = "Also shows the alliance tag next to your own name in the first person")
+    @ConfigurableBoolean(location = "alliances.general.showalliancetagonself")
+    @IncludeInAccordion("general")
+    private static boolean showAllianceTagOnSelf = false;
+
    public AllianceNameDecorator() {
       LifestealUtilsEvents.PLAYER_NAME_RENDER.register(this::onPlayerNameRender);
    }
 
    private void onPlayerNameRender(PlayerNameRenderEvent event) {
-      if (!Config.isEnableAlliances() || !Config.isAllianceNamePrefixEnabled()) {
+       if (!enableAlliances || !allianceNamePrefixEnabled) {
          return;
       }
       String playerName = event.getPlayerName();
@@ -25,7 +60,7 @@ public final class AllianceNameDecorator {
          return;
       }
 
-      if (!Config.isShowAllianceTagOnSelf()) {
+       if (!showAllianceTagOnSelf) {
          Minecraft client = Minecraft.getInstance();
          if (client.player != null && playerName.equals(client.player.getName().getString())) {
             return;
@@ -43,7 +78,7 @@ public final class AllianceNameDecorator {
          return;
       }
 
-      MutableComponent coloredName = Config.isAllianceNameColorEnabled()
+       MutableComponent coloredName = allianceNameColorEnabled
               ? colorOnlyPlayerName(event.getModifiedDisplayName(), playerName, decor.nameColor)
               : event.getModifiedDisplayName().copy();
       Component prefix = parsePrefix(decor.prefixText, decor.prefixFallback, decor.prefixColor);
