+++ apps/lsu/src/main/java/dev/candycup/lifestealutils/features/alliances/AllianceService.java
@@ -1,6 +1,7 @@
 package dev.candycup.lifestealutils.features.alliances;
 
-import dev.candycup.lifestealutils.Config;
+
+import dev.candycup.lifestealutils.gaia.GaiaConsentController;
 import net.minecraft.client.Minecraft;
 
 import java.util.ArrayList;
@@ -222,7 +223,7 @@ public final class AllianceService {
       record.clientId = AllianceIdGenerator.newClientId();
       record.order = localAlliances.size();
       record.canEdit = true;
-      record.subscriptionPermission = Config.isGaiaAdvancedFeaturesEnabled() ? "ANYONE" : "MEMBERS";
+      record.subscriptionPermission = GaiaConsentController.isGaiaAdvancedFeaturesEnabled() ? "ANYONE" : "MEMBERS";
       record.published = false;
       record.source = "local";
       record.syncState = "LOCAL";
@@ -313,3 +314,5 @@ public final class AllianceService {
       }
    }
 }
+
+
