+++ apps/lsu/src/main/java/dev/candycup/lifestealutils/features/messages/PrivateMessageFormatter.java
@@ -1,9 +1,13 @@
 package dev.candycup.lifestealutils.features.messages;
 
-import dev.candycup.lifestealutils.Config;
+import dev.candycup.configura.serial.SerialEntry;
+import dev.candycup.lifestealutils.config.configurables.ConfigurableBoolean;
+import dev.candycup.lifestealutils.config.configurables.ConfigurableMinimessage;
 import dev.candycup.lifestealutils.event.LifestealUtilsEvents;
 import dev.candycup.lifestealutils.event.LifestealUtilsEvents.ChatMessageReceivedEvent;
 import dev.candycup.lifestealutils.interapi.MessagingUtils;
+import lombok.Getter;
+import lombok.Setter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -15,6 +19,19 @@ import java.util.regex.Matcher;
  */
 public class PrivateMessageFormatter {
    private static final Logger LOGGER = LoggerFactory.getLogger("lifestealutils/pm");
+    public static final String DEFAULT_FORMAT = "<light_purple><bold>{{direction}}</bold> {{sender}}</light_purple> <white>\u27A1 {{message}}</white>";
+
+    @Getter
+    @Setter
+    @SerialEntry(comment = "Whether to enable custom private message formatting")
+    @ConfigurableBoolean(location = "customization.messages.pmformatenabled")
+    private static boolean enablePmFormat = false;
+
+    @Getter
+    @Setter
+    @SerialEntry(comment = "Customize the format of private messages (/msg, /r)")
+    @ConfigurableMinimessage(location = "customization.messages.pmformat")
+    private static String pmFormat = DEFAULT_FORMAT;
 
    public PrivateMessageFormatter() {
       LifestealUtilsEvents.CHAT_MESSAGE_RECEIVED.register(event -> {
@@ -26,7 +43,7 @@ public class PrivateMessageFormatter {
    }
 
    public boolean isEnabled() {
-      return Config.isEnablePmFormat();
+       return enablePmFormat;
    }
 
    public void onChatMessageReceived(ChatMessageReceivedEvent event) {
@@ -41,9 +58,7 @@ public class PrivateMessageFormatter {
       String sender = MessagingUtils.escapeMiniMessageTags(matcher.group(2));
       String message = MessagingUtils.escapeMiniMessageTags(matcher.group(3));
 
-      String format = Config.getPmFormat() != null && !Config.getPmFormat().isBlank()
-              ? Config.getPmFormat()
-              : "<light_purple><bold>{{direction}}</bold> {{sender}}</light_purple> <white>➡ {{message}}</white>";
+       String format = pmFormat != null && !pmFormat.isBlank() ? pmFormat : DEFAULT_FORMAT;
 
       String formatted = format
               .replace("{{direction}}", direction)
