+++ apps/lsu/src/main/java/dev/candycup/lifestealutils/features/titlescreen/CustomSplashes.java
@@ -1,10 +1,13 @@
 package dev.candycup.lifestealutils.features.titlescreen;
 
-import dev.candycup.lifestealutils.Config;
+import dev.candycup.configura.serial.SerialEntry;
 import dev.candycup.lifestealutils.FeatureFlagController;
+import dev.candycup.lifestealutils.config.configurables.ConfigurableBoolean;
 import dev.candycup.lifestealutils.event.LifestealUtilsEvents;
 import dev.candycup.lifestealutils.event.LifestealUtilsEvents.SplashTextRequestEvent;
 import dev.candycup.lifestealutils.interapi.MessagingUtils;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,6 +21,12 @@ public final class CustomSplashes {
       add("<yellow>uhoh...</yellow>");
    }};
 
+    @Getter
+    @Setter
+    @SerialEntry(comment = "Whether to enable custom splashes on the title screen")
+    @ConfigurableBoolean(location = "customization.titlescreen.customSplashes")
+    private static boolean customSplashesEnabled = true;
+
    public CustomSplashes() {
       LifestealUtilsEvents.SPLASH_TEXT_REQUEST.register(event -> {
          if (!isEnabled()) {
@@ -28,7 +37,7 @@ public final class CustomSplashes {
    }
 
    public boolean isEnabled() {
-      return Config.isCustomSplashesEnabled();
+       return customSplashesEnabled;
    }
 
    public void onSplashTextRequest(SplashTextRequestEvent event) {
